Move contents to auth package
Move auth classes to new package.
