+++ openAPE/server/src/main/java/org/openape/server/auth/AuthConfigFactory.java
@@ -1,4 +1,4 @@
-package org.openape.server;
+package org.openape.server.auth;
 
 import org.pac4j.core.config.Config;
 import org.pac4j.core.config.ConfigFactory;
